+++ tools/cli/dev-bundle-bin-helpers.js
@@ -1,11 +1,64 @@
 var fs = require("fs");
 var path = require("path");
 var files = require("../fs/mini-files.js");
+var isWindows = process.platform === "win32";
+var extensions = isWindows ? [".cmd", ".exe"] : [""];
+var hasOwn = Object.prototype.hasOwnProperty;
 
 function getDevBundle() {
   return require("./dev-bundle.js");
 }
 exports.getDevBundle = getDevBundle;
+
+exports.getCommand = function (name, devBundleDir) {
+  var result = null;
+
+  // Strip leading and/or trailing whitespace.
+  name = name.replace(/^\s+|\s+$/g, "");
+
+  if (! isValidCommand(name, devBundleDir)) {
+    return result;
+  }
+
+  extensions.some(function (ext) {
+    var cmd = path.join(devBundleDir, "bin", name + ext);
+    try {
+      if (fs.statSync(cmd).isFile()) {
+        result = cmd;
+        return true;
+      }
+    } catch (e) {
+      return false;
+    }
+  });
+
+  return result;
+};
+
+function isValidCommand(name, devBundleDir) {
+  if (name === "node" ||
+      name === "npm") {
+    return true;
+  }
+
+  if (! name || name.charAt(0) === ".") {
+    // Disallow empty commands and commands that start with a period.
+    return false;
+  }
+
+  var meteorCommandsJsonPath =
+    path.join(devBundleDir, "bin", ".meteor-commands.json");
+
+  try {
+    var meteorCommands = require(meteorCommandsJsonPath);
+  } catch (e) {
+    return false;
+  }
+
+  // If `meteor <name>` is already a Meteor command, don't let anything in
+  // dev_bundle/bin override it.
+  return ! hasOwn.call(meteorCommands, name);
+}
 
 exports.getEnv = function (options) {
   var devBundle = options && options.devBundle;